add stg_align field in STG_MEMBERS
This patch adds the stg_align field in STG_MEMBERS and
the relative function, to record the dtype's alignment
passed by `!dir$ align XX`.

stg_align field is alloced in stg_alloc_base(), deleted in
stg_delete() and reallocated in stg_need(), which is the same
as stg_base field.

Yet it will not be cleared by stg_clear*() functions, which can
helps flang1 to store value during semant,
even if stg_base field is cleared.