[flang1]store the dtype alignment from pragma
flang1 uses get_type() to create new dtype.

Considering that flang1 creates one and only one dtype
when parsing user-defined new types, so flang1 can
handle the possible align pragma for this dtype
in get_type().

flang1 should store the alignment from flg.x[251]
to DTA, later it will updates the dtype alignment
when in alignment() or lower_put_datatype().

Note that this pragma should only be used to
this strucutre, so flang1 need to clear the flg.x[251]